+++ source/mover2/changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.5</span>**
+- Ignore files on all future scans if "Remove source files" is not selected
+- Add ability to force tasks to be created for all files tested regardless of other plugin processing requirements.
+
 **<span style="color:#56adda">0.0.4</span>**
 - Enabled support for v2 plugin executor